+++ async_upnp_client/client.py
@@ -638,10 +638,9 @@ class UpnpAction:
         query = ".//{{{0}}}{1}Response".format(service_type, self.name)
         response = xml.find(query, NS)
 
-        # Strip version number from service type to be more tolerant to incomplete namespace.
-        if response is None and ":" in service_type and self._non_strict:
-            service_type_funny = ":".join(service_type.split(":")[:-1])
-            query = ".//{{{0}}}{1}Response".format(service_type_funny, self.name)
+        # If no response was found, do a search ignoring namespaces when in non-strict mode.
+        if response is None and self._non_strict:
+            query = ".//{{*}}{0}Response".format(self.name)
             response = xml.find(query, NS)
 
         if response is None: